requirements.txt
aiohttp~=3.6.2
ujson~=3.0.0
discord.py~=1.5.0